+++ src/main.js
@@ -5,11 +5,10 @@ import App from './App'
 import router from './router'
 import ElementUI from 'element-ui';
 import 'element-ui/lib/theme-chalk/index.css';
-  import Login from './login'
+import Login from '@/components/Login'
 Vue.use(ElementUI);
 Vue.config.productionTip = false
 
-/* eslint-disable no-new */
 new Vue({
   el: '#app',
   router,
